+++ doc/conf.py
@@ -5,8 +5,8 @@
 
 import os
 import sys
-from pathlib import Path
 from datetime import date
+from pathlib import Path
 
 from intersphinx_registry import get_intersphinx_mapping
 from sphinx.config import is_serializable
@@ -63,7 +63,7 @@ numpydoc_xref_aliases = {
     "list": ":class:`list <python:list>`",
     "tuple": ":class:`tuple <python:tuple>`",
     "NibabelImageObject": "nibabel.spatialimages.SpatialImage",
-    "CalibrationObject": "mne.preprocessing.eyetracking.Calibration"
+    "CalibrationObject": "mne.preprocessing.eyetracking.Calibration",
 }
 numpydoc_xref_ignore = {
     # words